% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_4_blank_utils_check_names.R
\name{check_blank}
\alias{check_blank}
\title{Checks sample-column from raw data, if any blank is there from metadata}
\usage{
check_blank(data_pmol)
}
\arguments{
\item{data}{tibble. Data table containing LipidView and/or LipidXplorer data. Already merged with metadata.}
}
\value{
vector
}
\description{
Checks sample-column from raw data, if any blank is there from metadata
}
\examples{


}
