% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_utils_check_names.R
\name{check_sample_names}
\alias{check_sample_names}
\title{Compares sample names from LipidView and LipidXplorer. If names are not
identical, check_sample_names() will stop the process}
\usage{
check_sample_names(data1, data2)
}
\arguments{
\item{data}{tibble. Data table containing LipidView or LipidXplorer data}

\item{column}{string. Name of the sample column}

\item{ignore}{string. Pattern to ignore e.g. "kon" -> Ignores Kon A, B and C}
}
\value{
vector
}
\description{
Compares sample names from LipidView and LipidXplorer. If names are not
identical, check_sample_names() will stop the process
}
\examples{
get_samples(

}
