% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_fct_data-pipeline.R
\name{clean_lipidview_files}
\alias{clean_lipidview_files}
\title{Clean up of LipidView txt.files}
\usage{
clean_lipidview_files(raw_data)
}
\arguments{
\item{raw_data}{tibble}
}
\value{
tibble
}
\description{
Clean up of LipidView txt.files and turn it into a long format
 data frame. "class"-column will be created by the string inside the brackets
 next to the lipid species.
}
\examples{
list.files(path = system.file("extdata", package = "shinylipidcountr"),
           pattern = ".txt", full.names = TRUE) \%>\%
  read_lipidview_files() \%>\%
  clean_lipidview_files()
}
