% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_fct_data-pipeline.R
\name{clean_lipidxplorer_files}
\alias{clean_lipidxplorer_files}
\title{Clean up of LipidXplorer csv-files}
\usage{
clean_lipidxplorer_files(raw_data, clean_samples = FALSE)
}
\arguments{
\item{raw_data}{tibble}

\item{clean_samples}{logical. TRUE activates janitor::clean_names() for samples

* if TRUE: "Sample 1" -> "sample_1" or "SampleOne" -> "sample_one".
But "Sample1" -> "sample1"!

* if FALSE: Default. Sample name remains untouched.}
}
\value{
tibble
}
\description{
Clean up of LipidXplorer csv-files
}
\examples{
system.file("extdata", "lx-out.csv", package = "shinylipidcountr") \%>\%
  read_lipidxplorer_files() \%>\%
  clean_lipidxplorer_files()
}
