% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lipidview_part_1.R
\name{clean_txt_files}
\alias{clean_txt_files}
\title{Clean up of LipidView txt.files}
\usage{
clean_txt_files(files)
}
\arguments{
\item{files}{tibble}
}
\value{
tibble
}
\description{
Clean up of LipidView txt.files
}
