% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_utils_check_names.R
\name{get_samples}
\alias{get_samples}
\title{Subsets all samples from data without controls}
\usage{
get_samples(data, column = "sample", ignore = "kon")
}
\arguments{
\item{data}{tibble. Data table containing LipidView or LipidXplorer data}

\item{column}{string. Name of the sample column}

\item{ignore}{string. Pattern to ignore e.g. "kon" -> Ignores Kon A, B and C}
}
\value{
vector
}
\description{
Subsets all samples from data without controls
}
