% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_3_pmol_fct_data-pipeline.R
\name{intensity_to_pmol}
\alias{intensity_to_pmol}
\title{Calculate pmol values}
\usage{
intensity_to_pmol(raw_data)
}
\arguments{
\item{raw_data}{tibble. Comes after join_istd_means()}
}
\value{
tibble
}
\description{
intensity_to_pmol() calculates pmol-values from intensity of
endogenous lipid species, intensity-mean of the class-specific internal
standard and the internal standard concentration in pmol.
(standard_input / is_intensity) * intensity)
}
\examples{
join_metadata(test_raw_data, test_metadata) \%>\%
  join_istd_means() \%>\%
  intensity_to_pmol()
}
