% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_3_pmol_fct_data-pipeline.R
\name{join_istd_means}
\alias{join_istd_means}
\title{Join averaged IS intensities to samples by class-column}
\usage{
join_istd_means(raw_data)
}
\arguments{
\item{raw_data}{tibble. Joined raw and metadata table.}
}
\value{
tibble
}
\description{
join_istd_means() extracts all rows with "IS" in
'species'-column and calculates the mean of same lipidclass-standards.
Afterwards these intensity-means are joined by 'class' and 'sample' in column
'is_intensity'
}
\examples{
join_metadata(test_raw_data, test_metadata) \%>\%
  join_istd_means()
}
