% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_2_meta_fct_data-pipeline.R
\name{join_metadata}
\alias{join_metadata}
\title{Join metadata with raw data}
\usage{
join_metadata(raw_data, metadata)
}
\arguments{
\item{raw_data}{data.frame. Raw data containing class, species, sample and
intensity data}

\item{metadata}{data.frame. Metadata containing sample volume, standard input,
response factors, sample groups}
}
\value{
data.frame
}
\description{
Merges raw data and metadata in one long-formatted table. Joins are done by
column 'sample' and 'class'. Therefore the sample- and class-values must be
identical in raw data and metadata!
}
\examples{
join_metadata(test_raw_data, test_metadata)
}
