% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipid_class_order}
\alias{lipid_class_order}
\title{Vector with Excelsheet names}
\format{
A \code{vector} with 71 elements
}
\usage{
lipid_class_order
}
\description{
A \code{vector} containing strings to set the order of sheets in the
Excelworkbook.
}
\keyword{datasets}
