% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_fct_data-pipeline.R
\name{merge_with_lipidview}
\alias{merge_with_lipidview}
\title{Merge data frame from LipidXplorer and LipidView.}
\usage{
merge_with_lipidview(lx_data, lv_data)
}
\arguments{
\item{lx_data}{tibble. Generated before by read_lipidview_files() and clean_lipidview_files()}

\item{lv_data}{tibble. Generated before by read_lipidxplorer_files() and clean_lipidxplorer_files()}
}
\value{
tibble
}
\description{
Merge data frame from LipidXplorer and LipidView.
}
