% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_fct_data-pipeline.R
\name{read_txt_files}
\alias{read_txt_files}
\title{Read (and merge) named txt.files, generated by LV}
\usage{
read_txt_files(files)
}
\arguments{
\item{files}{character vector. Vector of path/name from textfiles}
}
\value{
tibble
}
\description{
Read (and merge) named txt.files, generated by LV
}
\examples{
read_txt_files(system.file("extdata",
                            "yymmdd_XtrIL_ExampleData_SM.txt",
                            package = "lipidcountr"))
}
