% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_metadata}
\alias{test_metadata}
\title{Data frame with experimental data from lipid extraction}
\format{
A data frame with 18 rows and 23 variables:
\describe{
  \item{class}{lipid class abbreviation for the internal standards}
  \item{rf_values}{response factors}
  \item{Sample 1, etc}{contains experimental data like extractions inputs}
  \item{row 'group'}{sample name to summarise replicates}
  \item{row 'blank'}{blank assignment for blank extraction, if multiple blanks were used}
  \item{row 'sample input'}{sample volume used for extraction, in microliter}
  \item{intensity}{peak intensity of precursor, in cps}
  \item{PC, SM, etc}{concentration of internal standards used for extraction, in picomoles}
}
}
\source{
system.file("extdata", package = "shinylipidcountr")
}
\usage{
test_metadata
}
\description{
A dataset containing internal standards with concentration per sample used for extraction
}
\keyword{datasets}
