% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_raw_data_lv}
\alias{test_raw_data_lv}
\title{Tibble with lipidomics data from LipidView}
\format{
A data frame with 6867 rows and 5 variables:
\describe{
  \item{class}{lipid class abbreviation}
  \item{species}{lipid species, IS x:y;z, IS = internal standard, x = c-atoms, y = double bonds, z = hydroxilation grade}
  \item{scan_name}{precursor ion scan}
  \item{sample}{sample name/number according to sample entry sheet}
  \item{intensity}{peak intensity, in cps}
}
}
\source{
system.file("extdata", package = "shinylipidcountr")
}
\usage{
test_raw_data_lv
}
\description{
A dataset containing intensities and species of several lipid classes.
}
\keyword{datasets}
