% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_raw_data_lx}
\alias{test_raw_data_lx}
\title{Tibble with lipidomics data from LipidXplorer}
\format{
A data frame with 10298 rows and 8 variables:
\describe{
  \item{class}{lipid class abbreviation}
  \item{species}{lipid species, IS x:y;z, IS = internal standard, x = c-atoms, y = double bonds, z = hydroxilation grade}
  \item{molecular_species}{molecular species, IS x1:y1_x2:y2, x1:y1 = fatty acid1, x2:y2 = fatty acid2}
  \item{lipid_category}{lipid category, IS = Internal standard, PC = Glycerophospholipid, SM = Sphingolipid, TG = Glycerolipid}
  \item{mass}{exact mass of lipid, in Da}
  \item{sample}{sample name/number according to sample entry sheet}
  \item{intensity}{peak intensity of precursor, in cps}
  \item{fa_intensity}{sum intensity of fatty acid fragments, in cps}
}
}
\source{
system.file("extdata", package = "shinylipidcountr")
}
\usage{
test_raw_data_lx
}
\description{
A dataset containing intensities and species of several lipid classes.
}
\keyword{datasets}
