% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_2_meta_utils_check_names.R
\name{check_raw_meta}
\alias{check_raw_meta}
\title{Compares sample names from metadata file and rawdata. If names are not
identical, check_raw_meta() will stop the process}
\usage{
check_raw_meta(raw, metadata, ignore = "")
}
\arguments{
\item{raw}{tibble. Data table containing LipidView and/or LipidXplorer data}

\item{metadata}{data.frame. Data table containing metadata}

\item{ignore}{string. Pattern to ignore e.g. "kon" -> Ignores Kon A, B and C}
}
\value{
warning- or errormessage if sample names dont match.
}
\description{
Compares sample names from metadata file and rawdata. If names are not
identical, check_raw_meta() will stop the process
}
\examples{
check_raw_meta(test_raw_data, test_metadata)
}
