% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_fct_lipidview_part_1.R
\name{clean_txt_files}
\alias{clean_txt_files}
\title{Clean up of LipidView txt.files}
\usage{
clean_txt_files(rawdata, clean_samples = TRUE)
}
\arguments{
\item{rawdata}{tibble}

\item{clean_samples}{logical. TRUE activates janitor::clean_names() for samples
* if TRUE: Default. "Sample 1" -> "sample_1" or "SampleOne" -> "sample_one".
But "Sample1" -> "sample1"!

* if FALSE: Sample name remains untouched.}
}
\value{
tibble
}
\description{
Clean up of LipidView txt.files
}
