% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_fct_lipidview.R
\name{read_lipidview_files}
\alias{read_lipidview_files}
\title{Read, merge and clean txt.files, which are generated by LipidView}
\usage{
read_lipidview_files(path, clean_samples = FALSE, format = "long")
}
\arguments{
\item{path}{string. Path to your textfiles}

\item{clean_samples}{logical
* FALSE: Default. Sample name remains untouched.
* TRUE: activates janitor::clean_names() for samples}

\item{format}{string
* "long": Default. Prints the tibble in long-format (all samples in one col)
* "wide": Prints the tibble in wide-format (one sample, one col)}
}
\value{
tibble. Merged textfiles into one tibble
}
\description{
Read, merge and clean txt.files, which are generated by LipidView
}
\examples{
read_lipidview_files(system.file("extdata", package = "shinylipidcountr"))
}
