% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_1_raw_fct_lipidview_part_1.R
\name{read_lipidview_files}
\alias{read_lipidview_files}
\title{lipidview_part_1}
\usage{
read_lipidview_files(path, format = "long")
}
\arguments{
\item{path}{string. Path to your textfiles}

\item{format}{string
* "long": Default. Prints the tibble in long-format (all samples in one col)
* "wide": Prints the tibble in wide-format (one sample, one col)}
}
\value{
The return value, if any, from executing the function.

tibble. Merged textfiles into one tibble
}
\description{
A fct function
}
\examples{
read_lipidview_files(system.file("extdata", package = "shinylipidcountr"))
}
